requirements.txt
google.generativeai==0.8.3
python-dotenv==1.0.1
PyMuPDF==1.25.1
shapely==2.0.6
llama_index==0.12.16
llama-index-llms-gemini==0.4.10
weaviate==0.1.2
weaviate-client==4.10.4
llama-index-vector-stores-weaviate==1.3.1
langchain-openai==0.3.35
trulens_eval==1.4.0
trulens-providers-openai==1.4.0
trulens-apps-llamaindex==1.4.0